requirements.txt
Flask==3.0.0
PyJWT==2.8.0
joblib==1.3.2
numpy==1.24.3
SQLAlchemy==2.0.23
scikit-learn==1.3.2